Fixed a TODO and some errors
- Constructor now reads from Preferences as expected.
- DefaultSMSParseStrategy is standalone to avoid having to pass a context to instantiate it.
