+++ smslibrary/src/main/java/com/eis/smslibrary/SMSMessageParser.java
@@ -35,8 +35,9 @@ public class SMSMessageParser implements MessageHandler<String, String, SMSMessa
      * @param context The calling context
      */
     private SMSMessageParser(Context context) {
-        parseStrategy = new DefaultSMSMessageParseStrategy();
-        //TODO
+        SMSParseStrategy savedStrategy = getMessageParseStrategy(context);
+        if(savedStrategy == null) parseStrategy = new DefaultSMSParseStrategy();
+        else parseStrategy = savedStrategy;
     }
 
     /**
@@ -99,7 +100,7 @@ public class SMSMessageParser implements MessageHandler<String, String, SMSMessa
      * @param context the calling context
      */
     public void resetMessageParseStrategy(Context context){
-        this.parseStrategy = new DefaultSMSMessageParseStrategy();
+        this.parseStrategy = new DefaultSMSParseStrategy();
         PreferencesManager.removeValue(context, PREFERENCE_DEFAULT_STRATEGY);
     }
 
@@ -126,39 +127,6 @@ public class SMSMessageParser implements MessageHandler<String, String, SMSMessa
     public String parseData(@NonNull final SMSMessage message) {
         return parseStrategy.parseData(message);
     }
-
-    public class DefaultSMSMessageParseStrategy implements SMSParseStrategy {
-
-        protected static final String HIDDEN_CHARACTER = (char) 0x02 + "";
-
-        /**
-         * Parses sms data into a SMSMessage if possible
-         *
-         * @param channelPeer that sent the data
-         * @param channelData read from the channel
-         * @return the parsed SMSMessage if the string was correct, null otherwise
-         */
-        @Override
-        public SMSMessage parseMessage(@NonNull final SMSPeer channelPeer, @NonNull final String channelData) {
-            //First character of the content must be the hidden char
-            if (!channelData.startsWith(HIDDEN_CHARACTER))
-                return null;
-            String messageData = channelData.substring(1);
-            return new SMSMessage(channelPeer, messageData);
-        }
-
-        /**
-         * Parses SMSMessage into sms content data
-         *
-         * @param message from library
-         * @return the parsed sms content data ready to be sent
-         */
-        @Override
-        public String parseData(@NonNull final SMSMessage message) {
-            return HIDDEN_CHARACTER + message.getData();
-        }
-    }
-
 }
